separate functional slicing from classes, reorganize package structure a bit
