+++ tonic/__init__.py
@@ -1,3 +1,7 @@
 from .version import version as __version__
 from .utils import pad_tensors
 from . import datasets, transforms
+from .dataset import Dataset
+from .slicers import *
+from .sliced_dataset import SlicedDataset
+from .cached_dataset import CachedDataset
