+++ tonic/datasets/__init__.py
@@ -1,8 +1,6 @@
 from .asl_dvs import ASLDVS
 from .dvsgesture import DVSGesture
 from .davisdataset import DAVISDATA
-from .dataset import Dataset
-from .cached_dataset import CachedDataset
 from .hsd import SHD, SSC
 from .mvsec import MVSEC
 from .nmnist import NMNIST
