+++ tonic/datasets/dvsgesture.py
@@ -1,7 +1,7 @@
 import os
 import numpy as np
-from .dataset import Dataset
-from .download_utils import (
+from tonic.dataset import Dataset
+from tonic.download_utils import (
     check_integrity,
     download_and_extract_archive,
     extract_archive,
