+++ tonic/datasets/nmnist.py
@@ -2,8 +2,8 @@ import os
 import numpy as np
 from pathlib import Path
 
-from .dataset import Dataset
-from .download_utils import (
+from tonic.dataset import Dataset
+from tonic.download_utils import (
     check_integrity,
     download_and_extract_archive,
     extract_archive,
