+++ tonic/functional/__init__.py
@@ -5,6 +5,7 @@ from .drop_pixel import drop_pixel_numpy, identify_hot_pixel
 from .mix_ev_streams import mix_ev_streams_numpy
 from .refractory_period import refractory_period_numpy
 from .resize import spatial_resize_numpy
+from .slicing import slice_by_time, slice_by_time_bins, slice_by_event_count, slice_by_event_bins
 from .spatial_jitter import spatial_jitter_numpy
 from .time_jitter import time_jitter_numpy
 from .time_skew import time_skew_numpy
